backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
email-validator>=2.0.0
httpx>=0.24.0
aiofiles>=23.0.0
loguru>=0.7.0
